Fix bug when automatically bringing up a source file at a breakpoint.
If an exception is raised from a source file (or function) that is
not compiled with debug info (resulting in no file or lineno), the
EditorManager can't automatically open the next most available file (the
one from the next stack frame).

StackFramesBrowser now figures out which source file can be available
if the file for frame 0 isn't available.