Finished with alternate search directories for location source file.
GDB has its notion of where source files live. Sometimes this is
incorrect as the source may be in a different place after the executable
is released.

Seer's source search feature is a list of alternate directories to look
in if the path that GDB knows of doesn't exist.

Use the Config dialog to set them. Then save the config.

This is a global feature. That is, it applies to any and all executables
being debugged. (Perhaps I'll add project based session files to handle
per executable settings).