Finished Printpoints for now.
They always go to printf(). So the go to the console.
In the future, I can add an option to go to the "GDB output" tab.

Also, it may be possible to have a "Printpoint output" tab.
If the format has some magic text, the handleText() mechanism
may be able to intercept the output. For example, if the format
is:

    "^PrintPointFormat %d %d\n"

Of course this would work only if dprintf uses printf().