Notes for using RR and Seer.
It seems possible!  Probably need a new Debug tab called "RR" so that
I have a way to set the direction arrow (   <-- or -->  )  and hide
the "record" button. Because RR is just replaying the prevously recorded
session.  So "forward" and "back" buttons make sense. And no "record"
button make sense.